+++ CHANGELOG.md
@@ -11,7 +11,9 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Display omics variants - wts outliers (Fraser, Outrider)
 ### Changed
 - Unfreeze werkzeug again
-
+### Fixed
+- Sort "select default panels" dropdown menu options on case page
+- Show gene panel removed status on case page
 
 ## [4.84]
 ### Changed